Dynamic Footer for projects page
Dynamic footer that depends on global page size.

Before : When page contains only one row, page is smaller than viewport and footer is not at the bottom of viewport.

Current result : When the page's height is smaller than viewport's, position of footer is absolute. If it's the same or taller then the footer's position is relative.